python_chatbot/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asn1crypto==0.24.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.1.1
cryptography==2.1.4
cycler==0.10.0
decorator==4.1.2
discord.py @ git+https://github.com/Rapptz/discord.py@7cf3cd51a5d463ae5633ec2dd43d3e397b081876
frozenlist==1.3.3
idna==2.6
joblib==1.2.0
JPype1==1.4.1
keyring==10.6.0
keyrings.alt==3.0
konlpy==0.6.0
lxml==4.9.2
matplotlib==2.1.1
multidict==6.0.4
numpy==1.24.1
olefile==0.45.1
packaging==22.0
pandas==1.5.2
Pillow==5.1.0
pycrypto==2.6.1
PyGObject==3.26.1
PyMySQL==1.0.2
pyparsing==2.2.0
python-apt==1.6.4
python-dateutil==2.8.2
pytz==2022.7
pyxdg==0.25
requests==2.18.4
scikit-learn==1.2.0
scipy==1.9.3
SecretStorage==2.3.1
six==1.11.0
sklearn==0.0.post1
ssh-import-id==5.7
threadpoolctl==3.1.0
unattended-upgrades==0.1
urllib3==1.22
yarl==1.8.2
